Reverted as the clients are not ready for null values and other changes needs to be introduced to dev branch
Revert "Merge pull request #75 from mvougiou/dev"

This reverts commit e6fd8fee63f239e7f9596a20c73ad83c74032731, reversing
changes made to 707507e258f16b44be990e06f5891b4a1360c8ad.

+++ src/test/java/rcms/utilities/daqaggregator/datasource/MonitorManagerTest.java
@@ -90,7 +90,7 @@ public class MonitorManagerTest {
 	 */
 	private void testSnapshot(DAQ snapshot, long expectedTime, int expectedEvents) {
 		Assert.assertNotNull(snapshot);
-		Assert.assertEquals(expectedTime, snapshot.getLastUpdate().longValue());
+		Assert.assertEquals(expectedTime, snapshot.getLastUpdate());
 		Assert.assertEquals(expectedEvents, snapshot.getFedBuilderSummary().getSumEventsInRU());
 
 	}
